+++ cookbooks/aws-parallelcluster-environment/files/cloudwatch/write_cloudwatch_agent_json.py
@@ -26,7 +26,7 @@ def parse_args():
     parser.add_argument(
         "--node-role",
         required=True,
-        choices=["HeadNode", "ComputeFleet"],
+        choices=["HeadNode", "ComputeFleet", "LoginNode"],
         help="Role this node plays in the cluster " "(i.e., is it a compute node or the head node?)",
     )
     parser.add_argument("--scheduler", required=True, choices=["slurm", "awsbatch"], help="Scheduler")