+++ cookbooks/aws-parallelcluster-environment/recipes/config/mount_shared.rb
@@ -13,16 +13,9 @@
 
 return if on_docker?
 
-if node['cluster']['node_type'] == "ComputeFleet"
-  volume "mount /home" do
-    action :mount
-    shared_dir '/home'
-    device(lazy { "#{node['cluster']['head_node_private_ip']}:#{node['cluster']['head_node_home_path']}" })
-    fstype 'nfs'
-    options node['cluster']['nfs']['hard_mount_options']
-    retries 10
-    retry_delay 6
-  end
+case node['cluster']['node_type']
+when 'ComputeFleet'
+  include_recipe 'aws-parallelcluster-environment::mount_home'
 
   # Mount /opt/parallelcluster/shared over NFS
   volume "mount #{node['cluster']['shared_dir_compute']}" do
@@ -34,4 +27,22 @@ if node['cluster']['node_type'] == "ComputeFleet"
     retries 10
     retry_delay 6
   end
+
+when 'LoginNode'
+  include_recipe 'aws-parallelcluster-environment::mount_home'
+
+  # Mount /opt/parallelcluster/shared_login_nodes over NFS
+  volume "mount #{node['cluster']['shared_dir_login']}" do
+    action :mount
+    shared_dir node['cluster']['shared_dir_login']
+    device(lazy { "#{node['cluster']['head_node_private_ip']}:#{node['cluster']['shared_dir_login']}" })
+    fstype 'nfs'
+    options node['cluster']['nfs']['hard_mount_options']
+    retries 10
+    retry_delay 6
+  end
+when 'HeadNode'
+  Chef::Log.info("Nothing to mount in the HeadNode")
+else
+  raise "node_type must be HeadNode, LoginNode or ComputeFleet"
 end