+++ cookbooks/aws-parallelcluster-platform/resources/fetch_config.rb
@@ -10,8 +10,11 @@ property :update, [true, false],
 default_action :run
 
 action :run do
+  return if on_docker?
   Chef::Log.debug("Called fetch_config with update (#{new_resource.update})")
-  unless on_docker?
+
+  case node['cluster']['node_type']
+  when 'HeadNode'
     if new_resource.update
       Chef::Log.info("Backing up old configuration from (#{node['cluster']['cluster_config_path']}) to (#{node['cluster']['previous_cluster_config_path']})")
       ::FileUtils.cp_r(node['cluster']['cluster_config_path'], node['cluster']['previous_cluster_config_path'], remove_destination: true)
@@ -25,8 +28,21 @@ action :run do
       fetch_instance_type_data unless ::File.exist?(node['cluster']['instance_types_data_path'])
     end
 
-    # load cluster config into node object
-    load_cluster_config
+    # ensure config is shared also with login nodes
+    share_config_with_login_nodes
+
+    # load cluster config into a node object
+    load_cluster_config(node['cluster']['cluster_config_path'])
+  when 'ComputeFleet'
+    raise "Cluster config not found in #{node['cluster']['cluster_config_path']}" unless ::File.exist?(node['cluster']['cluster_config_path'])
+    # load cluster config into a node object
+    load_cluster_config(node['cluster']['cluster_config_path'])
+  when 'LoginNode'
+    raise "Cluster config not found in #{node['cluster']['login_cluster_config_path']}" unless ::File.exist?(node['cluster']['login_cluster_config_path'])
+    # load cluster config into a node object
+    load_cluster_config(node['cluster']['login_cluster_config_path'])
+  else
+    raise "node_type must be HeadNode, LoginNode or ComputeFleet"
   end
 end
 
@@ -66,6 +82,17 @@ action_class do # rubocop:disable Metrics/BlockLength
   def fetch_change_set
     # Copy change set file from S3 URI
     fetch_s3_object("copy_change_set_from_s3", node['cluster']['change_set_s3_key'], node['cluster']['change_set_path'])
+  end
+
+  def share_config_with_login_nodes
+    # Share cluster config with login nodes (only if they exist)
+    Chef::Log.info("Sharing cluster config with login nodes")
+    ::FileUtils.cp_r(node['cluster']['cluster_config_path'],
+                     node['cluster']['login_cluster_config_path'],
+                     remove_destination: true) unless !::File.exist?(node['cluster']['cluster_config_path'])
+    ::FileUtils.cp_r(node['cluster']['previous_cluster_config_path'],
+                     node['cluster']['login_previous_cluster_config_path'],
+                     remove_destination: true) unless !::File.exist?(node['cluster']['previous_cluster_config_path'])
   end
 
   def fetch_instance_type_data