+++ cookbooks/aws-parallelcluster-platform/test/controls/cluster_user_spec.rb
@@ -20,10 +20,10 @@ control 'tag:config_cluster_user' do
   end unless os_properties.redhat_on_docker?
 end
 
-control 'cluster_user_compute' do
-  title 'Check the cluster user configuration for compute node'
+control 'cluster_user_compute_and_login' do
+  title 'Check the cluster user configuration for compute node and login nodes'
 
-  only_if { !os_properties.on_docker? && instance.compute_node? }
+  only_if { !os_properties.on_docker? && (instance.compute_node? or instance.login_node?) }
 
   describe 'Check that cluster user exist'
   describe user('test_user') do