+++ cookbooks/aws-parallelcluster-slurm/recipes/config/config_login.rb
@@ -1,7 +1,10 @@
 # frozen_string_literal: true
 
 #
-# Copyright:: 2021-2023 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# Cookbook:: aws-parallelcluster-slurm
+# Recipe:: config_compute
+#
+# Copyright:: 2013-2023 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance with the
 # License. A copy of the License is located at
@@ -12,13 +15,7 @@
 # OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions and
 # limitations under the License.
 
-unless node['cluster']['scheduler'] == 'awsbatch'
-  case node['cluster']['node_type']
-  when 'HeadNode'
-    load_cluster_config(node['cluster']['cluster_config_path'])
-  else
-    raise "node_type must be HeadNode"
-  end
-end
+# TODO: rename, find a better name that include login nodes
+setup_munge_compute_node
 
-include_recipe 'aws-parallelcluster-slurm::update_computefleet_status' if node['cluster']['scheduler'] == 'slurm'
+include_recipe 'aws-parallelcluster-slurm::mount_slurm_dir'