+++ cookbooks/aws-parallelcluster-slurm/recipes/config/mount_slurm_dir.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
-#
-# Copyright:: 2021-2023 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# Copyright:: 2013-2023 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance with the
 # License. A copy of the License is located at
@@ -12,13 +11,13 @@
 # OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions and
 # limitations under the License.
 
-unless node['cluster']['scheduler'] == 'awsbatch'
-  case node['cluster']['node_type']
-  when 'HeadNode'
-    load_cluster_config(node['cluster']['cluster_config_path'])
-  else
-    raise "node_type must be HeadNode"
-  end
+# Mount /opt/slurm over NFS
+# Computemgtd config is under /opt/slurm/etc/pcluster; all compute nodes share a config
+mount "#{node['cluster']['slurm']['install_dir']}" do
+  device(lazy { "#{node['cluster']['head_node_private_ip']}:#{node['cluster']['slurm']['install_dir']}" })
+  fstype "nfs"
+  options node['cluster']['nfs']['hard_mount_options']
+  action %i(mount enable)
+  retries 10
+  retry_delay 6
 end
-
-include_recipe 'aws-parallelcluster-slurm::update_computefleet_status' if node['cluster']['scheduler'] == 'slurm'